liuxiaodong，2018年9月18日17点59分，创建主账号；删除主账号；修改主账号密码；修改主账号扩展信息

+++ uoo-core/uoo-personnel/src/main/java/cn/ffcs/uoo/core/personnel/entity/TbAcctOrg.java
@@ -26,17 +26,17 @@ public class TbAcctOrg extends Model<TbAcctOrg> {
      * 主账号与组织关系标识
      */
     @TableId("ACCT_ORG_ID")
-    private BigDecimal acctOrgId;
+    private long acctOrgId;
     /**
      * 主账号标识
      */
     @TableField("ACCT_ID")
-    private BigDecimal acctId;
+    private long acctId;
     /**
      * 组织标识
      */
     @TableField("ORG_ID")
-    private BigDecimal orgId;
+    private long orgId;
     /**
      * 关系类型 主归属、兼职、借调
      */
@@ -84,27 +84,27 @@ public class TbAcctOrg extends Model<TbAcctOrg> {
     private Date statusDate;
 
 
-    public BigDecimal getAcctOrgId() {
+    public long getAcctOrgId() {
         return acctOrgId;
     }
 
-    public void setAcctOrgId(BigDecimal acctOrgId) {
+    public void setAcctOrgId(long acctOrgId) {
         this.acctOrgId = acctOrgId;
     }
 
-    public BigDecimal getAcctId() {
+    public long getAcctId() {
         return acctId;
     }
 
-    public void setAcctId(BigDecimal acctId) {
+    public void setAcctId(long acctId) {
         this.acctId = acctId;
     }
 
-    public BigDecimal getOrgId() {
+    public long getOrgId() {
         return orgId;
     }
 
-    public void setOrgId(BigDecimal orgId) {
+    public void setOrgId(long orgId) {
         this.orgId = orgId;
     }
